+++ stac_fastapi/pgstac/tests/resources/test_item.py
@@ -82,6 +82,13 @@ async def test_create_item(app_client, load_test_data: Callable, load_test_colle
     get_item = Item.parse_obj(resp.json())
     assert in_item.dict(exclude={"links"}) == get_item.dict(exclude={"links"})
 
+    post_self_link = next(
+        (link for link in post_item.links if link.rel == "self"), None
+    )
+    get_self_link = next((link for link in get_item.links if link.rel == "self"), None)
+    assert post_self_link is not None and get_self_link is not None
+    assert post_self_link.href == get_self_link.href
+
 
 async def test_create_item_mismatched_collection_id(
     app_client, load_test_data: Callable, load_test_collection
@@ -141,6 +148,7 @@ async def test_update_item(
         f"/collections/{coll.id}/items/{item.id}", content=item.json()
     )
     assert resp.status_code == 200
+    put_item = Item.parse_obj(resp.json())
 
     resp = await app_client.get(f"/collections/{coll.id}/items/{item.id}")
     assert resp.status_code == 200
@@ -148,6 +156,11 @@ async def test_update_item(
     get_item = Item.parse_obj(resp.json())
     assert item.dict(exclude={"links"}) == get_item.dict(exclude={"links"})
     assert get_item.properties.description == "Update Test"
+
+    post_self_link = next((link for link in put_item.links if link.rel == "self"), None)
+    get_self_link = next((link for link in get_item.links if link.rel == "self"), None)
+    assert post_self_link is not None and get_self_link is not None
+    assert post_self_link.href == get_self_link.href
 
 
 async def test_update_item_mismatched_collection_id(